+++ R/MoEClust-package.R
@@ -56,7 +56,7 @@
 #' (comp <- MoE_compare(res1, res2, optimal.only=TRUE))
 #' 
 #' # Produce a plot for the optimal model                                                   
-#' plot(comp$optimal, what="gpairs")
+#' plot(comp$optimal, what="gpairs", show.dens=TRUE)
 #' 
 #' # Summarise its classification table, component parameters, and gating/expert networks
 #' summary(comp$optimal, classification=TRUE, parameters=TRUE, networks=TRUE)